`commands` function can return a promise (#67)
The app developer can provide a `commands` function that returns a promise. When the user invokes the context menu it will appear only once the promise is resolved. This allows the app developer to involve async operations when calculating the available commands.

The returned promise must resolve with an array of commands. Note: the array itself must not contain any promise.

The cxtmenu API remains fully compatible. The `commands` function still can return an array of commands.

The README is extended.